// lib/api/central/requests.ts
  export interface TenantPayload {
    name: string
    slug?: string
    database?: string
    domain: string
    status?: TenantStatus
    plan_id?: string | null
    billing_cycle: BillingCycle
    owner_email: string
    owner_name: string
  }